Remove unnecessary subset definitions
A test playlist defines the subset(s) that a test belongs to. If no
subsets are defined, the test defaults to running for all subsets. In
some cases, a test playlist is defined with SE80 / SE90 / SE100 / SE110
- all the test subsets except for Panama & Valhalla. Many test cases
which are valid for SE80 / SE90 / SE100 / SE110 are also valid for
Panama and Valhalla builds. This commit removes the subset tag from the
tests that are valid for all six cases.

Signed-off-by: Talia McCormick <Talia.McCormick@ibm.com>